Add files via upload
This ino file is for a wired version from the Attiny85 controller using 2 wire serial communications.  Do not excess 4800 baud due to faster speeds cause a lot of garbage in the serial stream.  Hook up the TX pin from the 85 to the TX pin on the Pi.  Hook up the RX pin from the 85 to the RX pin on the Pi.